update setup script

+++ setup.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
-files="vimrc zshrc zprofile bashrc profile gitconfig gitignore_global"
+files="vimrc zshrc zprofile bashrc profile gitconfig gitignore_global condarc"
 repositories="CSM152A CSM148 CSM148.3 randallwc.github.io"
 
 echo "making simlinks in home dir"
@@ -14,29 +14,27 @@ do
     fi
 done
 
-if [[ ! -d ~/Github/ ]]
+# if github dir does not exist
+if [[ ! -d ~/Github/ ]] && [[ "$1" == "--github" ]]
 then
-    if [[ "$1" == "--github" ]]
+    echo "making GitHub in home directory"
+    mkdir ~/GitHub/
+    cd ~/GitHub/
+    git --version >/dev/null 2>&1
+    if [[ $? -ne 0 ]]
     then
-        echo "making GitHub in home directory"
-        mkdir ~/GitHub/
-        cd ~/GitHub/
-        git --version >/dev/null 2>&1
+        echo "git not installed"
+        exit 1
+    fi
+    for repo in $repositories
+    do
+        echo "cloning $repo ..."
+        git clone git@github.com:randallwc/$repo
         if [[ $? -ne 0 ]]
         then
-            echo "git not installed"
+            echo "error occured in git clone"
+            rmdir ~/GitHub
             exit 1
         fi
-        for repo in $repositories
-        do
-            echo "cloning $repo ..."
-            git clone git@github.com:randallwc/$repo
-            if [[ $? -ne 0 ]]
-            then
-                echo "error occured in git clone"
-                rmdir ~/GitHub
-                exit 1
-            fi
-        done
-    fi
+    done
 fi
